windows-server/requirements.txt
fastapi
uvicorn
pydantic
requests
Pillow
qrcode
zeroconf
websockets
